+++ lib/std/os/linux.zig
@@ -2054,7 +2054,7 @@ pub fn eventfd(count: u32, flags: u32) usize {
     return syscall2(.eventfd2, count, flags);
 }
 
-pub fn timerfd_create(clockid: u32, flags: TFD) usize {
+pub fn timerfd_create(clockid: usize, flags: TFD) usize {
     return syscall2(
         .timerfd_create,
         clockid,
@@ -4555,7 +4555,7 @@ pub const clockid_t = enum(u32) {
 // to a panic, an enum is introduced which only allows the values
 // that actually work.
 pub const CLOCK_ID = clock_id;
-pub const clock_id = enum(u32) {
+pub const clock_id = enum(usize) {
     REALTIME = 0,
     MONOTONIC = 1,
     _,
